backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
python-jose==3.3.0
passlib==1.7.4
cryptography==42.0.0
httpx==0.27.0
click==8.1.7
bcrypt==4.1.2